More planet generation work.
Now either everything in generating, place-holdered or todo'ed.

+++ Pulsar4X/Pulsar4X.Lib/Entities/StarSystem/Star.cs
@@ -39,7 +39,7 @@ namespace Pulsar4X.Entities
     {
         public BindingList<Planet> Planets { get; set; }
         public double Age { get; set; }
-        public uint Temperature { get; set; } // Effective ("Photosphere") temperature in Degrees C.
+        public double Temperature { get; set; } // Effective ("Photosphere") temperature in Degrees C.
         public float Luminosity { get; set; }
         public string Class { get; set; } 
         public double EcoSphereRadius { get; set; } // Average echo sphere. TODO: change this to include min and max radius from GetHabitableZone
